BC-1471 - rename file
