+++ apps/server/src/modules/user/mapper/index.ts
@@ -1,1 +1,1 @@
-export * from './ResolvedUser.mapper';
+export * from './resolved-user.mapper';
